pyproject.toml
[project]
name = "local-agents"
version = "0.1.0"
description = "Multi-agent AI system with orchestrator architecture"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "openai>=1.60.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "jinja2>=3.1.6",
    "duckduckgo-search>=7.0.0",
    "fastmcp>=2.0.0",
    "websockets>=14.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "pyobjc-framework-avfoundation>=12.1",
    "soundfile>=0.13.1",
    "sounddevice>=0.5.5",
    "scipy>=1.17.0",
    "huggingface-hub>=0.34.0",
    "qwen-tts>=0.1.1",
    "torchaudio>=2.8.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.20.0",
    "mlx-audio>=0.2.0",
]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "app", "channels"]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
